Moved "run_command" from "src/main.c" to "src/exectute.c".
